+++ tkviews/core/__init__.py
@@ -1,3 +1,4 @@
-'''Core package'''
+"""Core package"""
 
 from .node import TkNode
+from .rendering import TkRenderingContext