+++ tkviews/widgets/tests/ttk_tests.py
@@ -4,9 +4,10 @@ from pytest import mark
 from pyviews.core import XmlAttr
 from pyviews.pipes import call_set_attr
 
-from tkviews import ttk
+from tkviews.widgets import ttk
 from tkviews.core.rendering import TkRenderingContext
-from tkviews.ttk import TtkStyle, theme_use, setup_value_setter, configure, apply_style_attributes
+from tkviews.widgets.ttk import TtkStyle, theme_use, setup_value_setter, configure, \
+    apply_style_attributes
 
 
 class TtkStyleTests: